Populated TopoJSON object with data
Added the population density data as a property to the TopoJSON object that is shown on the map. Further seperation between fetching data and drawing. Moved most of the functions in map.js and moved the intialization of the map in to it's own function. Added back the main file to make execution easier to follow